requirements.txt
alembic==1.12.0
aniso8601==9.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
future==0.18.3
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
pycryptodome==3.18.0
python-dotenv==0.21.1
python-jose-cryptodome==1.3.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
